working scala

+++ ceasar_cypher.scala
@@ -1,9 +1,9 @@
-import Cryptor
+import cryptor.Cryptor
 
 object CeasarCypher {
 
   def main(args: Array[String]) {
-    val cryptor = new Cryptor(args(0), args(1))
+    val cryptor = new Cryptor(args(0), args(1).toInt)
     cryptor.encrypt
   }
 }
